fix(frontend): update tally popup copy
